Convert header wrappers to vanilla-extract
