+++ apps/store/src/components/ForeverPage/ForeverPage.tsx
@@ -5,7 +5,7 @@ import { FormEventHandler, useState } from 'react'
 import { Button, mq, Space, theme } from 'ui'
 import { useGlobalBanner } from '@/components/GlobalBanner/useGlobalBanner'
 import { GridLayout } from '@/components/GridLayout/GridLayout'
-import { HEADER_HEIGHT_DESKTOP, HEADER_HEIGHT_MOBILE } from '@/components/Header/Header'
+import { HEADER_HEIGHT_MOBILE, HEADER_HEIGHT_DESKTOP } from '@/components/Header/Header.constants'
 import { TextField } from '@/components/TextField/TextField'
 import { TextWithLink } from '@/components/TextWithLink'
 import { useShopSession } from '@/services/shopSession/ShopSessionContext'
