+++ apps/store/src/components/Header/Header.constants.ts
@@ -1,3 +1,8 @@
+import { theme } from 'ui'
+
 export const MENU_BAR_HEIGHT_MOBILE = '3rem'
 export const MENU_BAR_HEIGHT_DESKTOP = '4rem'
 export const MENU_BAR_HEIGHT_PX = 64
+
+export const HEADER_HEIGHT_MOBILE = `calc(${MENU_BAR_HEIGHT_MOBILE} + ${theme.space.xs})`
+export const HEADER_HEIGHT_DESKTOP = `calc(${MENU_BAR_HEIGHT_DESKTOP} + ${theme.space.xs})`
