+++ apps/store/src/features/carDealership/CarBuyerValidationPage.tsx
@@ -3,7 +3,7 @@ import styled from '@emotion/styled'
 import type { TextSprinkles } from 'ui/src/components/Text/Text.css'
 import { Button, Heading, HedvigLogo, Space, Text, mq, theme } from 'ui'
 import { GridLayout } from '@/components/GridLayout/GridLayout'
-import { HEADER_HEIGHT_DESKTOP, HEADER_HEIGHT_MOBILE } from '@/components/Header/Header'
+import { HEADER_HEIGHT_MOBILE, HEADER_HEIGHT_DESKTOP } from '@/components/Header/Header.constants'
 import { AttentionCard, CampaignCard } from '@/components/InfoCard/InfoCard'
 import { InputCarRegistrationNumber } from '@/components/InputCarRegistrationNumber/InputCarRegistrationNumber'
 import { InputSelect } from '@/components/InputSelect/InputSelect'
